requirements.txt
Bottleneck==1.3.4
brotlipy==0.7.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.4
cryptography==37.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2
django-import-export==2.5.0
django-simpleui==2021.8.1
et-xmlfile==1.1.0
idna==3.3
MarkupPy==1.14
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.8.1
numpy==1.21.5
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
pip==21.2.4
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.4
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
setuptools==61.2.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
urllib3==1.26.9
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
